flask_auth/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Mako==1.3.0
MarkupSafe==2.1.3
phonenumbers==8.13.27
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.32.1
rpds-py==0.16.2
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.2
